+++ tests/integration/test_spans.py
@@ -72,6 +72,7 @@ def test_span_extraction(
             "attributes": {"txn_key": 123},
         },
     ]
+    event["contexts"]["replay"] = {"replay_id": "4c79f60c11214eb38604f4ae0781bfb2"}
     end = datetime.now(timezone.utc) - timedelta(seconds=1)
     duration = timedelta(milliseconds=500)
     start = end - duration
@@ -135,6 +136,10 @@ def test_span_extraction(
             "sentry.op": {"type": "string", "value": "http"},
             "sentry.origin": {"type": "string", "value": "manual"},
             "sentry.platform": {"type": "string", "value": "other"},
+            "sentry.replay_id": {
+                "type": "string",
+                "value": "4c79f60c11214eb38604f4ae0781bfb2",
+            },
             "sentry.sdk.name": {"type": "string", "value": "raven-node"},
             "sentry.sdk.version": {"type": "string", "value": "2.6.3"},
             "sentry.status": {"type": "string", "value": "ok"},
@@ -150,6 +155,12 @@ def test_span_extraction(
             "sentry.user.geo.subregion": {"type": "string", "value": "155"},
             "sentry.user.id": {"type": "string", "value": user_id},
             "sentry.user.ip": {"type": "string", "value": "192.168.0.1"},
+            "user.geo.city": {"type": "string", "value": "Vienna"},
+            "user.geo.country_code": {"type": "string", "value": "AT"},
+            "user.geo.region": {"type": "string", "value": "Austria"},
+            "user.geo.subdivision": {"type": "string", "value": "Vienna"},
+            "user.id": {"type": "string", "value": user_id},
+            "user.ip_address": {"type": "string", "value": "192.168.0.1"},
             "sentry.description": {
                 "type": "string",
                 "value": "GET /api/0/organizations/?member=1",
@@ -217,6 +228,10 @@ def test_span_extraction(
             "sentry.op": {"type": "string", "value": "hi"},
             "sentry.origin": {"type": "string", "value": "manual"},
             "sentry.platform": {"type": "string", "value": "other"},
+            "sentry.replay_id": {
+                "type": "string",
+                "value": "4c79f60c11214eb38604f4ae0781bfb2",
+            },
             "sentry.sdk.name": {"type": "string", "value": "raven-node"},
             "sentry.sdk.version": {"type": "string", "value": "2.6.3"},
             "sentry.segment.id": {"type": "string", "value": "968cff94913ebb07"},
@@ -225,6 +240,7 @@ def test_span_extraction(
             "sentry.trace.status": {"type": "string", "value": "ok"},
             "sentry.transaction.op": {"type": "string", "value": "hi"},
             "sentry.transaction": {"type": "string", "value": "hi"},
+            "sentry.user": {"type": "string", "value": f"id:{user_id}"},
             "sentry.user.geo.city": {"type": "string", "value": "Vienna"},
             "sentry.user.geo.country_code": {"type": "string", "value": "AT"},
             "sentry.user.geo.region": {"type": "string", "value": "Austria"},
@@ -232,7 +248,12 @@ def test_span_extraction(
             "sentry.user.geo.subregion": {"type": "string", "value": "155"},
             "sentry.user.id": {"type": "string", "value": user_id},
             "sentry.user.ip": {"type": "string", "value": "192.168.0.1"},
-            "sentry.user": {"type": "string", "value": f"id:{user_id}"},
+            "user.geo.city": {"type": "string", "value": "Vienna"},
+            "user.geo.country_code": {"type": "string", "value": "AT"},
+            "user.geo.region": {"type": "string", "value": "Austria"},
+            "user.geo.subdivision": {"type": "string", "value": "Vienna"},
+            "user.id": {"type": "string", "value": user_id},
+            "user.ip_address": {"type": "string", "value": "192.168.0.1"},
             "sentry.dsc.project_id": {"type": "string", "value": "42"},
             "sentry.dsc.trace_id": {
                 "type": "string",
@@ -488,20 +509,14 @@ def envelope_with_transaction_and_spans(start: datetime, end: datetime) -> Envel
     return envelope
 
 
-@pytest.mark.parametrize("measurements_conversion", ["direct", "smart"])
 def test_span_ingestion_with_performance_scores(
-    mini_sentry, relay_with_processing, spans_consumer, measurements_conversion
+    mini_sentry, relay_with_processing, spans_consumer
 ):
     spans_consumer = spans_consumer()
     relay = relay_with_processing()
 
     project_id = 42
     project_config = mini_sentry.add_full_project_config(project_id)
-    if measurements_conversion == "smart":
-        project_config["config"]["features"] = [
-            "projects:relay-measurements-smart-conversion"
-        ]
-
     project_config["config"]["performanceScore"] = {
         "profiles": [
             {
@@ -607,29 +622,6 @@ def test_span_ingestion_with_performance_scores(
     # endpoint might overtake envelope
     spans.sort(key=lambda msg: msg["span_id"])
 
-    if measurements_conversion == "smart":
-        measurements1 = {
-            "browser.web_vital.cls.value": 100.0,
-            "browser.web_vital.fcp.value": 200.0,
-            "fid": 300.0,
-            "browser.web_vital.lcp.value": 400.0,
-            "browser.web_vital.ttfb.value": 500.0,
-        }
-        measurements2 = {
-            "browser.web_vital.inp.value": 100.0,
-        }
-    else:
-        measurements1 = {
-            "cls": 100.0,
-            "fcp": 200.0,
-            "fid": 300.0,
-            "lcp": 400.0,
-            "ttfb": 500.0,
-        }
-        measurements2 = {
-            "inp": 100.0,
-        }
-
     expected_scores = [
         {
             "score.fcp": 0.14999972769539766,
@@ -647,15 +639,19 @@ def test_span_ingestion_with_performance_scores(
             "score.weight.fid": 0.3,
             "score.weight.lcp": 0.3,
             "score.weight.ttfb": 0.0,
+            "browser.web_vital.cls.value": 100.0,
+            "browser.web_vital.fcp.value": 200.0,
+            "fid": 300.0,
+            "browser.web_vital.lcp.value": 400.0,
+            "browser.web_vital.ttfb.value": 500.0,
             "score.cls": 0.0,
-            **measurements1,
         },
         {
+            "browser.web_vital.inp.value": 100.0,
             "score.inp": 0.9948129113413748,
             "score.ratio.inp": 0.9948129113413748,
             "score.total": 0.9948129113413748,
             "score.weight.inp": 1.0,
-            **measurements2,
         },
     ]
 
@@ -1156,11 +1152,21 @@ def test_scrubs_ip_addresses(
 
     child_span = spans_consumer.get_span()
 
+    assert child_span["_meta"]["attributes"]["user.email"] == {
+        "": {"len": 15, "rem": [["@email", "s", 0, 7]]}
+    }
     assert child_span["_meta"]["attributes"]["sentry.user.email"] == {
         "": {"len": 15, "rem": [["@email", "s", 0, 7]]}
     }
 
     if scrub_ip_addresses:
+        assert child_span["attributes"]["user.ip_address"] is None
+        assert child_span["_meta"]["attributes"]["user.ip_address"] == {
+            "": {
+                "len": 9,
+                "rem": [["@ip:replace", "s", 0, 4], ["@anything:remove", "x"]],
+            }
+        }
         assert child_span["attributes"]["sentry.user.ip"] is None
         assert child_span["_meta"]["attributes"]["sentry.user.ip"] == {
             "": {
@@ -1169,14 +1175,18 @@ def test_scrubs_ip_addresses(
             }
         }
     else:
+        assert child_span["attributes"]["user.ip_address"]["value"] == "127.0.0.1"
+        assert "user.ip_address" not in child_span["_meta"]["attributes"]
         assert child_span["attributes"]["sentry.user.ip"]["value"] == "127.0.0.1"
         assert "sentry.user.ip" not in child_span["_meta"]["attributes"]
 
     parent_span = spans_consumer.get_span()
 
     if scrub_ip_addresses:
+        assert "user.ip_address" not in parent_span["attributes"]
         assert "sentry.user.ip" not in parent_span["attributes"]
     else:
+        assert parent_span["attributes"]["user.ip_address"]["value"] == "127.0.0.1"
         assert parent_span["attributes"]["sentry.user.ip"]["value"] == "127.0.0.1"
 
     spans_consumer.assert_empty()
@@ -1254,6 +1264,8 @@ def test_spans_dsc_normalization(
     spans_consumer = spans_consumer()
     ts = datetime.now(timezone.utc)
     event = make_transaction({"event_id": "cbf6960622e14a45abc1f03b2055b186"})
+    # Large data for testing if DSC attributes are trimmed from trace context
+    event["contexts"]["trace"]["data"] = {"big_content": "1" * 10000}
     event["spans"] = [
         {
             "trace_id": "a0fa8803753e40fd8124b21eeb2986b5",